requirements.txt
glfw
pyopengl
Pyopengl accelerate
pyrr
pillow